+++ src/main/java/org/hankki/hankkiserver/api/auth/service/AuthService.java
@@ -8,6 +8,10 @@ import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.hankki.hankkiserver.api.auth.service.response.UserInfoResponse;
 import org.hankki.hankkiserver.api.auth.service.response.UserLoginResponse;
+import org.hankki.hankkiserver.api.user.service.UserFinder;
+import org.hankki.hankkiserver.api.user.service.UserInfoFinder;
+import org.hankki.hankkiserver.api.user.service.UserInfoUpdater;
+import org.hankki.hankkiserver.api.user.service.UserUpdater;
 import org.hankki.hankkiserver.auth.jwt.JwtProvider;
 import org.hankki.hankkiserver.auth.jwt.Token;
 import org.hankki.hankkiserver.domain.user.model.Platform;
